refactor: add upload component dir
