+++ JS/views/components/upload/UploadHeaderView.js
