+++ JS/views/components/upload/UploadPhotoListView.js
