+++ JS/views/components/upload/UploadPhotoView.js
